fix: convert useCountup to an CountUpAnimation component
